+++ frontend/src/pages/LogRegPage.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import { Container } from "react-bootstrap";
 
-import backgroundImage from "/bg.avif";
-import LoginForm from "../components/LoginForm";
 import RegisterForm from "../components/RegisterForm";
+import LoginForm from "../components/LoginForm";
+import backgroundImage from "/bg.avif";
 import "../styles/FlipCard.css";
 
 export default function FlipCard() {
@@ -34,7 +34,7 @@ export default function FlipCard() {
             <div className='flip-card-front'>
               <LoginForm />
               <p
-                className='text-center text-warning d-block mt-4 flip-button'
+                className='text-center text-warning d-block mt-4 text-decoration-underline'
                 onClick={() => setIsFlipped(true)}>
                 No account? Click to register
               </p>
@@ -44,7 +44,7 @@ export default function FlipCard() {
             <div className='flip-card-back'>
               <RegisterForm />
               <p
-                className='text-center text-warning d-block mt-3 flip-button'
+                className='text-center text-warning d-block mt-3 text-decoration-underline'
                 onClick={() => setIsFlipped(false)}>
                 Have an account? Click to login
               </p>